Add ESlint, TypeScript configurations for backend. Typing backend code in progress
