--- backend/controllers/testing.js
@@ -1,12 +1,0 @@
-const router = require("express").Router();
-const Note = require("../models/note");
-const User = require("../models/user");
-
-router.post("/reset", async (request, response) => {
-  await Note.deleteMany({});
-  await User.deleteMany({});
-
-  response.status(204).end();
-});
-
-module.exports = router;
