+++ backend/tests/average.test.ts
@@ -1,0 +1,23 @@
+import { average } from "../utils/for_testing";
+import { test, describe } from "node:test";
+import assert from "node:assert";
+
+
+// const { test, describe } = require("node:test");
+// const assert = require("node:assert");
+
+// const average = require("../utils/for_testing").average;
+
+describe("average", () => {
+	test("of one value is the value itself", () => {
+		assert.strictEqual(average([1]), 1);
+	});
+
+	test("of many values is calculated right", () => {
+		assert.strictEqual(average([1, 2, 3, 4, 5, 6]), 3.5);
+	});
+
+	test("of empty array is zero", () => {
+		assert.strictEqual(average([]), 0);
+	});
+});
