+++ backend/tests/reverse.test.ts
@@ -1,0 +1,26 @@
+import { reverse } from "../utils/for_testing";
+import { test } from "node:test";
+import assert from "node:assert";
+
+// const { test } = require("node:test");
+// const assert = require("node:assert");
+
+// const reverse = require("../utils/for_testing").reverse;
+
+test("reverse of a", () => {
+	const result = reverse("a");
+
+	assert.strictEqual(result, "a");
+});
+
+test("reverse of react", () => {
+	const result = reverse("react");
+
+	assert.strictEqual(result, "tcaer");
+});
+
+test("reverse of saippuakauppias", () => {
+	const result = reverse("saippuakauppias");
+
+	assert.strictEqual(result, "saippuakauppias");
+});
