+++ backend/tests/user_api.test.ts
@@ -1,0 +1,81 @@
+import bcrypt from "bcrypt";
+import { test, describe, after, beforeEach } from "node:test";
+import assert from "node:assert";
+import supertest from "supertest";
+import mongoose from "mongoose";
+import helper from "./test_helper";
+import app from "../app";
+import User from "../models/user";
+
+
+// const bcrypt = require("bcrypt");
+// const { test, beforeEach, after, describe } = require("node:test");
+// const assert = require("node:assert");
+// const supertest = require("supertest");
+// const mongoose = require("mongoose");
+// const helper = require("./test_helper");
+// const app = require("../app");
+const api = supertest(app);
+
+// const User = require("../models/user");
+
+describe("when there is initally one user in db", () => {
+	beforeEach(async () => {
+		await User.deleteMany({});
+
+		const passwordHash = await bcrypt.hash("sekret", 10);
+		const user = new User({
+			username: "root",
+			passwordHash,
+		});
+
+		await user.save();
+	});
+
+	test("creation succeeds with a fresh username", async () => {
+		const usersAtStart = await helper.usersInDb();
+
+		const newUser = {
+			username: "cinamonbunn",
+			name: "Chloe",
+			password: "salainen",
+		};
+
+		await api
+			.post("/api/users")
+			.send(newUser)
+			.expect(201)
+			.expect("Content-Type", /application\/json/);
+
+		const usersAtEnd = await helper.usersInDb();
+		assert.strictEqual(usersAtEnd.length, usersAtStart.length + 1);
+
+		const usernames = usersAtEnd.map((user) => user.username);
+		assert(usernames.includes(newUser.username));
+	});
+
+	test("creation fails with proper statuscode and message if username already taken", async () => {
+		const usersAtStart = await helper.usersInDb();
+
+		const newUser = {
+			username: "root",
+			name: "Superuser",
+			password: "salainen",
+		};
+
+		const result = await api
+			.post("/api/users")
+			.send(newUser)
+			.expect(400)
+			.expect("Content-Type", /application\/json/);
+
+		const usersAtEnd = await helper.usersInDb();
+		assert(result.body.error.includes("expected `username` to be unique"));
+
+		assert.strictEqual(usersAtEnd.length, usersAtStart.length);
+	});
+});
+
+after(async () => {
+	await mongoose.connection.close();
+});
