--- backend/utils/for_testing.js
@@ -1,16 +1,0 @@
-const reverse = (string) => {
-  return string.split("").reverse().join("");
-};
-
-const average = (array) => {
-  const reducer = (sum, item) => {
-    return sum + item;
-  };
-
-  return array.length === 0 ? 0 : array.reduce(reducer, 0) / array.length;
-};
-
-module.exports = {
-  reverse,
-  average,
-};
